Split DID into DIDSet and DIDDelete
